requirements.txt
# Core ML/DL libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# API and data collection
requests>=2.31.0
python-dotenv>=1.0.0

# Visualization and logging
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
tensorboard>=2.13.0

# Evaluation metrics
scipy>=1.11.0

